% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMprep_fxns.R
\name{bg.sel}
\alias{bg.sel}
\title{Background Point Selection}
\usage{
bg.sel(x)
}
\arguments{
\item{param1}{x = output from the bg.bb function.}
}
\value{
The output is a dataframe with columns 'x', 'y' corresponding to longitude and latitude.
}
\description{
This function selects background points from the area specified in the bg.bb function.
}
\details{
This function randomly selects 10000 background points from the area specified in the bg.bb function.
}
