% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMprep_fxns.R
\name{occ.prep}
\alias{occ.prep}
\title{Occurrence Preparation}
\usage{
occ.prep(x, y)
}
\arguments{
\item{param1}{y = the distance (in km) at which you would like to rarefy points}
}
\value{
The output is a dataframe.
}
\description{
This function prepares occurrence data for species distribution modeling
}
\details{
This function deletes duplicates and NAs from occurrence data. It also rarefies data by a user-specified distance.
}
