% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMprep_fxns.R
\name{partition}
\alias{partition}
\title{Data partitioning}
\usage{
partition(x, y)
}
\arguments{
\item{param1}{x = output from the occ.prep function, or a data frame with occurrence points and column names 'name', 'latitude', 'longitude'.}

\item{param2}{y = output from the bg.sel function.}
}
\value{
The output is two numeric vectors with numbers 1-4, one vector each for occurrences and background.
}
\description{
This function partitions both occurrence and background points into groups.
}
\details{
This function partitions occurrence data and background points into 4 groups to be used in k-fold jackknifing.
}
