% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMprep_fxns.R
\name{vis.results}
\alias{vis.results}
\title{Visualization of Results}
\usage{
vis.results(x, y)
}
\arguments{
\item{param1}{x = modeling output.}

\item{param2}{y = raster stack of environmental variables.}
}
\value{
The output is a text file with the variable importance for each model, a csv with all of the evaluation data for each model, a jpg with the OERs plotted for each model, a jpg with the AUCs plotted for each model, a jpg with the variable repsonses for the best model, and a jpg and tif file with the best model's prediction.
}
\description{
This function plots the evaluation metrics, variable response, and final prediction.
}
\details{
This function plots the ommission error rate and AUC values for all models. It selects the model with the lowest OER and plots the variable response curves and final prediction map for that model.
}
